Reorganize concurrent tests
